+++ src/main/java/org/truffleruby/core/thread/ThreadNodes.java
@@ -174,7 +174,7 @@ public abstract class ThreadNodes {
 
             final SafepointAction safepointAction = (thread1, currentNode) -> {
                 final Backtrace backtrace = getContext().getCallStack().getBacktrace(this, omit);
-                backtraceLocationsMemo.set(backtrace.getBacktraceLocations(length, this));
+                backtraceLocationsMemo.set(backtrace.getBacktraceLocations(getContext(), length, this));
             };
 
             getContext()