+++ CHANGELOG.md
@@ -10,6 +10,7 @@ New features:
 * `foreign_object.class` now calls `getMetaObject()` (except for Java classes, same as before).
 * Add basic support for Linux ARM64.
 * `foreign_object.name = value` will now call `Interoplibrary#writeMember("name", value)` instead of `invokeMember("name=", value)`.
+* Always show the Ruby core library files in backtraces (#1414).
 
 Bug fixes: